refactor: découpler les 4 composants React critiques
- TabTreeEditor : extraction du hook useTabTreeEditor (état + handlers),
  TabEditRow, GroupEditRow et MoveTabDropdown dans leurs propres fichiers
- DomainRulesPage : extraction de RuleDetailPopover (60 LOC de JSX inline),
  BulkActionsBar et RulesEmptyState comme composants locaux
- ImportSessionsWizard : déplacement de SessionRow, ConflictSessionRow,
  SessionDiffView et GroupDiffView dans SessionImportRows.tsx partagé
- ExportSessionsWizard : suppression des formatters dupliqués (formatDate,
  getTabCount, getGroupLabel, getTabLabel)
- sessionUtils : ajout de formatSessionDateShort, getSessionGroupLabel,
  getSessionTabLabel — maintenant partagés entre Import et Export
- options.tsx : extraction de useDeepLinking (hook), OptionsHeader et
  OptionsFooter (composants expanded/collapsed) dans OptionsLayout/

0 régression — 367 tests passés, pnpm compile propre.
