+++ src/components/Core/TabTree/TabTreeEditor.tsx
@@ -1,23 +1,23 @@
-import React, { useState, useEffect, useCallback } from 'react';
+import React from 'react';
 import {
   Flex,
   Text,
   Box,
   IconButton,
-  TextField,
   Button,
   ScrollArea,
-  DropdownMenu,
   AlertDialog,
 } from '@radix-ui/themes';
-import { ChevronUp, ChevronDown, Pencil, Trash2, ArrowUpRight } from 'lucide-react';
+import { ChevronUp, ChevronDown, Pencil, Trash2 } from 'lucide-react';
 import { getMessage } from '../../../utils/i18n';
 import { extractDomain } from './tabTreeUtils';
 import { TabRowBase } from './TabRowBase';
 import { GroupRowBase } from './GroupRowBase';
-import { ChromeColorPicker } from './ChromeColorPicker';
-import type { ChromeGroupColor } from './tabTreeTypes';
-import type { Session, SavedTab, SavedTabGroup } from '../../../types/session';
+import { TabEditRow } from './TabEditRow';
+import { GroupEditRow } from './GroupEditRow';
+import { MoveTabDropdown } from './MoveTabDropdown';
+import { useTabTreeEditor } from './useTabTreeEditor';
+import type { Session } from '../../../types/session';
 import styles from './TabTreeEditor.module.css';
 
 export interface TabTreeEditorProps {
@@ -29,212 +29,33 @@ export interface TabTreeEditorProps {
   maxHeight?: number | string;
 }
 
-type AlertDialogState =
-  | { type: 'delete_last_tab'; tabId: string; groupId: string }
-  | { type: 'delete_group'; groupId: string; groupTitle: string; tabCount: number }
-  | null;
-
 export function TabTreeEditor({ session, onSessionChange, maxHeight }: TabTreeEditorProps) {
-  const [expandedGroups, setExpandedGroups] = useState<Set<string>>(
-    () => new Set(session.groups.map((g) => g.id))
-  );
-  const [editingItemId, setEditingItemId] = useState<string | null>(null);
-  const [editUrl, setEditUrl] = useState('');
-  const [editGroupName, setEditGroupName] = useState('');
-  const [editGroupColor, setEditGroupColor] = useState<ChromeGroupColor>('grey');
-  const [urlError, setUrlError] = useState<string | null>(null);
-  const [alertDialog, setAlertDialog] = useState<AlertDialogState>(null);
-
-  // Auto-expand newly added groups
-  useEffect(() => {
-    setExpandedGroups((prev) => {
-      const next = new Set(prev);
-      let changed = false;
-      for (const group of session.groups) {
-        if (!next.has(group.id)) {
-          next.add(group.id);
-          changed = true;
-        }
-      }
-      return changed ? next : prev;
-    });
-  }, [session.groups]);
-
-  const now = () => new Date().toISOString();
-
-  const toggleGroup = useCallback((groupId: string) => {
-    setExpandedGroups((prev) => {
-      const next = new Set(prev);
-      if (next.has(groupId)) next.delete(groupId);
-      else next.add(groupId);
-      return next;
-    });
-  }, []);
-
-  function openTabEdit(tab: SavedTab) {
-    setEditingItemId(tab.id);
-    setEditUrl(tab.url);
-    setUrlError(null);
-  }
-
-  function openGroupEdit(group: SavedTabGroup) {
-    setEditingItemId(group.id);
-    setEditGroupName(group.title);
-    setEditGroupColor(group.color);
-  }
-
-  function cancelEdit() {
-    setEditingItemId(null);
-    setUrlError(null);
-  }
-
-  function saveTabEdit(tabId: string) {
-    let url = editUrl.trim();
-    // Auto-prefix https:// if no scheme is present
-    if (url && !/^[a-zA-Z][a-zA-Z\d+\-.]*:\/\//.test(url)) {
-      url = 'https://' + url;
-    }
-    try {
-      new URL(url);
-    } catch {
-      setUrlError(getMessage('tabEditorUrlInvalid'));
-      return;
-    }
-    onSessionChange({
-      ...session,
-      ungroupedTabs: session.ungroupedTabs.map((t) => (t.id === tabId ? { ...t, url } : t)),
-      groups: session.groups.map((g) => ({
-        ...g,
-        tabs: g.tabs.map((t) => (t.id === tabId ? { ...t, url } : t)),
-      })),
-      updatedAt: now(),
-    });
-    setEditingItemId(null);
-    setUrlError(null);
-  }
-
-  function saveGroupEdit(groupId: string) {
-    onSessionChange({
-      ...session,
-      groups: session.groups.map((g) =>
-        g.id === groupId ? { ...g, title: editGroupName, color: editGroupColor } : g
-      ),
-      updatedAt: now(),
-    });
-    setEditingItemId(null);
-  }
-
-  function moveTabInContext(tabId: string, direction: 'up' | 'down', contextGroupId: string | null) {
-    if (contextGroupId === null) {
-      const tabs = [...session.ungroupedTabs];
-      const idx = tabs.findIndex((t) => t.id === tabId);
-      if (idx === -1) return;
-      const newIdx = direction === 'up' ? idx - 1 : idx + 1;
-      if (newIdx < 0 || newIdx >= tabs.length) return;
-      [tabs[idx], tabs[newIdx]] = [tabs[newIdx], tabs[idx]];
-      onSessionChange({ ...session, ungroupedTabs: tabs, updatedAt: now() });
-    } else {
-      const groups = session.groups.map((g) => {
-        if (g.id !== contextGroupId) return g;
-        const tabs = [...g.tabs];
-        const idx = tabs.findIndex((t) => t.id === tabId);
-        if (idx === -1) return g;
-        const newIdx = direction === 'up' ? idx - 1 : idx + 1;
-        if (newIdx < 0 || newIdx >= tabs.length) return g;
-        [tabs[idx], tabs[newIdx]] = [tabs[newIdx], tabs[idx]];
-        return { ...g, tabs };
-      });
-      onSessionChange({ ...session, groups, updatedAt: now() });
-    }
-  }
-
-  function moveGroupInList(groupId: string, direction: 'up' | 'down') {
-    const groups = [...session.groups];
-    const idx = groups.findIndex((g) => g.id === groupId);
-    if (idx === -1) return;
-    const newIdx = direction === 'up' ? idx - 1 : idx + 1;
-    if (newIdx < 0 || newIdx >= groups.length) return;
-    [groups[idx], groups[newIdx]] = [groups[newIdx], groups[idx]];
-    onSessionChange({ ...session, groups, updatedAt: now() });
-  }
-
-  function moveTabToGroup(
-    tabId: string,
-    sourceGroupId: string | null,
-    targetGroupId: string | null
-  ) {
-    let tab: SavedTab | undefined;
-    let newUngrouped = session.ungroupedTabs;
-    let newGroups = session.groups;
-
-    if (sourceGroupId === null) {
-      tab = session.ungroupedTabs.find((t) => t.id === tabId);
-      newUngrouped = session.ungroupedTabs.filter((t) => t.id !== tabId);
-    } else {
-      const group = session.groups.find((g) => g.id === sourceGroupId);
-      tab = group?.tabs.find((t) => t.id === tabId);
-      newGroups = session.groups.map((g) =>
-        g.id === sourceGroupId ? { ...g, tabs: g.tabs.filter((t) => t.id !== tabId) } : g
-      );
-    }
-
-    if (!tab) return;
-
-    if (targetGroupId === null) {
-      newUngrouped = [...newUngrouped, tab];
-    } else {
-      newGroups = newGroups.map((g) =>
-        g.id === targetGroupId ? { ...g, tabs: [...g.tabs, tab!] } : g
-      );
-    }
-
-    onSessionChange({ ...session, ungroupedTabs: newUngrouped, groups: newGroups, updatedAt: now() });
-  }
-
-  function handleDeleteTab(tabId: string, sourceGroupId: string | null) {
-    if (sourceGroupId !== null) {
-      const group = session.groups.find((g) => g.id === sourceGroupId);
-      if (group && group.tabs.length === 1) {
-        setAlertDialog({ type: 'delete_last_tab', tabId, groupId: sourceGroupId });
-        return;
-      }
-    }
-    performDeleteTab(tabId, sourceGroupId, false);
-  }
-
-  function performDeleteTab(
-    tabId: string,
-    sourceGroupId: string | null,
-    deleteEmptyGroup: boolean
-  ) {
-    const newUngrouped = session.ungroupedTabs.filter((t) => t.id !== tabId);
-    let newGroups = session.groups.map((g) => ({
-      ...g,
-      tabs: g.tabs.filter((t) => t.id !== tabId),
-    }));
-    if (deleteEmptyGroup && sourceGroupId) {
-      newGroups = newGroups.filter((g) => g.id !== sourceGroupId);
-    }
-    onSessionChange({ ...session, ungroupedTabs: newUngrouped, groups: newGroups, updatedAt: now() });
-    setAlertDialog(null);
-  }
-
-  function handleDeleteGroup(groupId: string, action: 'delete_tabs' | 'ungroup_tabs') {
-    const group = session.groups.find((g) => g.id === groupId);
-    if (!group) return;
-    const newUngrouped =
-      action === 'ungroup_tabs'
-        ? [...session.ungroupedTabs, ...group.tabs]
-        : session.ungroupedTabs;
-    const newGroups = session.groups.filter((g) => g.id !== groupId);
-    onSessionChange({
-      ...session,
-      ungroupedTabs: newUngrouped,
-      groups: newGroups,
-      updatedAt: now(),
-    });
-    setAlertDialog(null);
-  }
+  const {
+    expandedGroups,
+    editingItemId,
+    editUrl,
+    setEditUrl,
+    setUrlError,
+    editGroupName,
+    setEditGroupName,
+    editGroupColor,
+    setEditGroupColor,
+    urlError,
+    alertDialog,
+    setAlertDialog,
+    toggleGroup,
+    openTabEdit,
+    openGroupEdit,
+    cancelEdit,
+    saveTabEdit,
+    saveGroupEdit,
+    moveTabInContext,
+    moveGroupInList,
+    moveTabToGroup,
+    handleDeleteTab,
+    performDeleteTab,
+    handleDeleteGroup,
+  } = useTabTreeEditor(session, onSessionChange);
 
   const content = (
     <Box className={styles.container} role="list">
@@ -604,164 +425,4 @@ export function TabTreeEditor({ session, onSessionChange, maxHeight }: TabTreeEd
   }
 
   return content;
-}
-
-/* ─── Sub-components ──────────────────────────────────────────────────────── */
-
-interface TabEditRowProps {
-  url: string;
-  error: string | null;
-  level: number;
-  onChange: (url: string) => void;
-  onSave: () => void;
-  onCancel: () => void;
-}
-
-function TabEditRow({ url, error, level, onChange, onSave, onCancel }: TabEditRowProps) {
-  return (
-    <Box
-      style={{
-        paddingLeft: (level - 1) * 20 + 8,
-        paddingRight: 'var(--space-2)',
-        paddingTop: 'var(--space-2)',
-        paddingBottom: 'var(--space-2)',
-      }}
-    >
-      <Flex direction="column" gap="2">
-        <Flex align="center" gap="2">
-          <Text size="1" color="gray" style={{ flexShrink: 0 }}>
-            {getMessage('tabEditorUrlLabel')}:
-          </Text>
-          <TextField.Root
-            value={url}
-            onChange={(e: React.ChangeEvent<HTMLInputElement>) => onChange(e.target.value)}
-            onKeyDown={(e: React.KeyboardEvent) => {
-              if (e.key === 'Enter') onSave();
-              if (e.key === 'Escape') onCancel();
-            }}
-            size="1"
-            style={{ flex: 1 }}
-            placeholder="https://example.com"
-            aria-label={getMessage('tabEditorUrlLabel')}
-            autoFocus
-          />
-        </Flex>
-        {error && (
-          <Text size="1" className={styles.urlError}>
-            {error}
-          </Text>
-        )}
-        <Flex gap="2" justify="end">
-          <Button size="1" variant="soft" color="gray" onClick={onCancel}>
-            {getMessage('cancel')}
-          </Button>
-          <Button size="1" onClick={onSave}>
-            {getMessage('save')}
-          </Button>
-        </Flex>
-      </Flex>
-    </Box>
-  );
-}
-
-interface GroupEditRowProps {
-  name: string;
-  color: ChromeGroupColor;
-  level: number;
-  onNameChange: (name: string) => void;
-  onColorChange: (color: ChromeGroupColor) => void;
-  onSave: () => void;
-  onCancel: () => void;
-}
-
-function GroupEditRow({
-  name,
-  color,
-  level,
-  onNameChange,
-  onColorChange,
-  onSave,
-  onCancel,
-}: GroupEditRowProps) {
-  return (
-    <Box
-      style={{
-        paddingLeft: (level - 1) * 20 + 8,
-        paddingRight: 'var(--space-2)',
-        paddingTop: 'var(--space-2)',
-        paddingBottom: 'var(--space-2)',
-      }}
-    >
-      <Flex direction="column" gap="2">
-        <Flex align="center" gap="2">
-          <TextField.Root
-            value={name}
-            onChange={(e: React.ChangeEvent<HTMLInputElement>) => onNameChange(e.target.value)}
-            onKeyDown={(e: React.KeyboardEvent) => {
-              if (e.key === 'Enter') onSave();
-              if (e.key === 'Escape') onCancel();
-            }}
-            size="1"
-            style={{ flex: 1 }}
-            placeholder={getMessage('tabEditorGroupNameLabel')}
-            aria-label={getMessage('tabEditorGroupNameLabel')}
-            autoFocus
-          />
-          <ChromeColorPicker value={color} onChange={onColorChange} />
-        </Flex>
-        <Flex gap="2" justify="end">
-          <Button size="1" variant="soft" color="gray" onClick={onCancel}>
-            {getMessage('cancel')}
-          </Button>
-          <Button size="1" onClick={onSave}>
-            {getMessage('save')}
-          </Button>
-        </Flex>
-      </Flex>
-    </Box>
-  );
-}
-
-interface MoveTabDropdownProps {
-  currentGroupId: string | null;
-  groups: SavedTabGroup[];
-  onMove: (targetGroupId: string | null) => void;
-}
-
-function MoveTabDropdown({ currentGroupId, groups, onMove }: MoveTabDropdownProps) {
-  const otherGroups = groups.filter((g) => g.id !== currentGroupId);
-  const canUngroup = currentGroupId !== null;
-
-  if (otherGroups.length === 0 && !canUngroup) return null;
-
-  return (
-    <DropdownMenu.Root>
-      <DropdownMenu.Trigger>
-        <IconButton
-          size="1"
-          variant="ghost"
-          color="gray"
-          aria-label={getMessage('tabEditorMoveTab')}
-          title={getMessage('tabEditorMoveTab')}
-        >
-          <ArrowUpRight size={12} aria-hidden="true" />
-        </IconButton>
-      </DropdownMenu.Trigger>
-      <DropdownMenu.Content>
-        {otherGroups.map((g) => (
-          <DropdownMenu.Item key={g.id} onClick={() => onMove(g.id)}>
-            {g.title}
-          </DropdownMenu.Item>
-        ))}
-        {canUngroup && (
-          <>
-            {otherGroups.length > 0 && <DropdownMenu.Separator />}
-            <DropdownMenu.Item onClick={() => onMove(null)}>
-              {getMessage('tabEditorUngroupedTabs')}
-            </DropdownMenu.Item>
-          </>
-        )}
-      </DropdownMenu.Content>
-    </DropdownMenu.Root>
-  );
 }
