+++ src/components/UI/ImportExportPage/ImportSessionsWizard.tsx
@@ -1,9 +1,9 @@
 import React, { useState, useCallback, useEffect, useRef, useMemo } from 'react';
 import {
-  Dialog, Flex, Button, Checkbox, Text, Separator, Badge, Box,
-  ScrollArea, TextArea, SegmentedControl, Popover, Callout
+  Dialog, Flex, Button, Text, Separator, Box,
+  ScrollArea, TextArea, SegmentedControl, Callout
 } from '@radix-ui/themes';
-import { Upload, FileUp, CheckCircle, AlertTriangle, XCircle, Eye } from 'lucide-react';
+import { Upload, FileUp, CheckCircle, AlertTriangle, XCircle } from 'lucide-react';
 import { z } from 'zod';
 import { SessionsTheme } from '../../Form/themes';
 import { getMessage } from '../../../utils/i18n';
@@ -12,12 +12,10 @@ import { sessionsArraySchema } from '../../../schemas/session';
 import {
   classifyImportedSessions,
   type SessionClassification,
-  type ConflictingSession,
-  type SessionDiff,
-  type GroupDiff,
 } from '../../../utils/sessionClassification';
 import { generateUUID } from '../../../utils/utils';
 import { loadSessions, saveSessions } from '../../../utils/sessionStorage';
+import { SessionRow, ConflictSessionRow } from './SessionImportRows';
 import type { Session } from '../../../types/session';
 
 type ConflictMode = 'overwrite' | 'duplicate' | 'ignore';
@@ -44,14 +42,12 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
   const [fileName, setFileName] = useState<string | null>(null);
   const [existingSessions, setExistingSessions] = useState<Session[]>([]);
 
-  // Step 1 state
   const [classification, setClassification] = useState<SessionClassification | null>(null);
   const [newSessionSelectedIds, setNewSessionSelectedIds] = useState<Set<string>>(new Set());
   const [conflictMode, setConflictMode] = useState<ConflictMode>('overwrite');
 
   const fileInputRef = useRef<HTMLInputElement>(null);
 
-  // Load existing sessions and reset state when dialog opens
   useEffect(() => {
     if (open) {
       loadSessions().then(loaded => setExistingSessions(loaded));
@@ -83,11 +79,7 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
       if (error instanceof SyntaxError) {
         setParseError(getMessage('invalidJson'));
       } else if (error instanceof z.ZodError) {
-        const messages = error.issues.map(issue => {
-          const path = issue.path.join('.');
-          return `${path}: ${issue.message}`;
-        });
-        setParseError(messages.join('\n'));
+        setParseError(error.issues.map(issue => `${issue.path.join('.')}: ${issue.message}`).join('\n'));
       } else {
         setParseError(getMessage('errorImportInvalidStructure'));
       }
@@ -114,9 +106,7 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
     e.preventDefault();
     setIsDragOver(false);
     const file = e.dataTransfer.files[0];
-    if (file && file.name.endsWith('.json')) {
-      handleFileRead(file);
-    }
+    if (file && file.name.endsWith('.json')) handleFileRead(file);
   }, [handleFileRead]);
 
   const handleDragOver = useCallback((e: React.DragEvent) => {
@@ -124,20 +114,15 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
     setIsDragOver(true);
   }, []);
 
-  const handleDragLeave = useCallback(() => {
-    setIsDragOver(false);
-  }, []);
+  const handleDragLeave = useCallback(() => setIsDragOver(false), []);
 
-  const handleBrowse = useCallback(() => {
-    fileInputRef.current?.click();
-  }, []);
+  const handleBrowse = useCallback(() => fileInputRef.current?.click(), []);
 
   const handleFileSelect = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
     const file = e.target.files?.[0];
     if (file) handleFileRead(file);
   }, [handleFileRead]);
 
-  // Transition to step 1: classify sessions
   const goToStep1 = useCallback(() => {
     if (!parsedSessions) return;
     const result = classifyImportedSessions(parsedSessions, existingSessions);
@@ -149,16 +134,12 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
   const toggleNewSession = useCallback((id: string) => {
     setNewSessionSelectedIds(prev => {
       const next = new Set(prev);
-      if (next.has(id)) {
-        next.delete(id);
-      } else {
-        next.add(id);
-      }
+      if (next.has(id)) next.delete(id);
+      else next.add(id);
       return next;
     });
   }, []);
 
-  // Compute import count for step 1
   const importCount = useMemo(() => {
     if (!classification) return 0;
     const newCount = classification.newSessions.filter(s => newSessionSelectedIds.has(s.id)).length;
@@ -175,7 +156,6 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
 
     const takenNames = new Set<string>(existingSessions.map(s => s.name.toLowerCase()));
 
-    // Add selected new sessions
     for (const session of classification.newSessions) {
       if (newSessionSelectedIds.has(session.id)) {
         updatedSessions.push(session);
@@ -184,7 +164,6 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
       }
     }
 
-    // Handle conflicts
     for (const conflict of classification.conflictingSessions) {
       if (conflictMode === 'overwrite') {
         const idx = updatedSessions.findIndex(
@@ -200,7 +179,6 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
         updatedSessions.push({ ...conflict.imported, id: generateUUID(), name: uniqueName });
         added++;
       }
-      // 'ignore': do nothing
     }
 
     await saveSessions(updatedSessions);
@@ -210,23 +188,6 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
       getMessage('importSessionsNotificationMessage', [String(added), String(overwritten)]),
     );
   }, [classification, existingSessions, newSessionSelectedIds, conflictMode, onOpenChange]);
-
-  const formatDate = (iso: string) => {
-    try {
-      return new Date(iso).toLocaleDateString(undefined, { dateStyle: 'medium' });
-    } catch {
-      return iso;
-    }
-  };
-
-  const getTabCount = (session: Session) =>
-    session.groups.reduce((sum, g) => sum + g.tabs.length, 0) + session.ungroupedTabs.length;
-
-  const getGroupLabel = (count: number) =>
-    count === 1 ? getMessage('sessionGroupOne') : getMessage('sessionGroupCount').replace('$1', String(count));
-
-  const getTabLabel = (count: number) =>
-    count === 1 ? getMessage('sessionTabOne') : getMessage('sessionTabCount').replace('$1', String(count));
 
   const hasConflicts = (classification?.conflictingSessions.length ?? 0) > 0;
 
@@ -254,12 +215,8 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
                 onValueChange={(v: string) => setSourceMode(v as 'file' | 'text')}
                 size="2"
               >
-                <SegmentedControl.Item value="file">
-                  {getMessage('sourceFile')}
-                </SegmentedControl.Item>
-                <SegmentedControl.Item value="text">
-                  {getMessage('sourceText')}
-                </SegmentedControl.Item>
+                <SegmentedControl.Item value="file">{getMessage('sourceFile')}</SegmentedControl.Item>
+                <SegmentedControl.Item value="text">{getMessage('sourceText')}</SegmentedControl.Item>
               </SegmentedControl.Root>
 
               <Box mt="3">
@@ -292,15 +249,11 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
                       onClick={handleBrowse}
                     >
                       <FileUp size={32} style={{ color: 'var(--gray-9)' }} aria-hidden="true" />
-                      <Text size="2" color="gray">
-                        {getMessage('dragDropZone')}
-                      </Text>
+                      <Text size="2" color="gray">{getMessage('dragDropZone')}</Text>
                       <Button variant="soft" size="1" onClick={(e: React.MouseEvent) => { e.stopPropagation(); handleBrowse(); }}>
                         {getMessage('browse')}
                       </Button>
-                      {fileName && (
-                        <Text size="1" color="gray" mt="1">{fileName}</Text>
-                      )}
+                      {fileName && <Text size="1" color="gray" mt="1">{fileName}</Text>}
                     </Flex>
                   </>
                 )}
@@ -316,23 +269,16 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
                 )}
               </Box>
 
-              {/* Validation feedback */}
               {parseError && (
                 <Callout.Root color="red" variant="soft" mt="3">
-                  <Callout.Icon>
-                    <XCircle size={16} />
-                  </Callout.Icon>
-                  <Callout.Text style={{ whiteSpace: 'pre-wrap', fontSize: 12 }}>
-                    {parseError}
-                  </Callout.Text>
+                  <Callout.Icon><XCircle size={16} /></Callout.Icon>
+                  <Callout.Text style={{ whiteSpace: 'pre-wrap', fontSize: 12 }}>{parseError}</Callout.Text>
                 </Callout.Root>
               )}
 
               {parsedSessions && (
                 <Callout.Root color="green" variant="soft" mt="3">
-                  <Callout.Icon>
-                    <CheckCircle size={16} />
-                  </Callout.Icon>
+                  <Callout.Icon><CheckCircle size={16} /></Callout.Icon>
                   <Callout.Text>
                     {getMessage('sessionsFoundCount').replace('{count}', String(parsedSessions.length))}
                   </Callout.Text>
@@ -346,7 +292,6 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
             <Box mt="4">
               <ScrollArea type="auto" scrollbars="vertical" style={{ maxHeight: '50vh' }}>
                 <Flex direction="column" gap="3" pr="3">
-                  {/* New sessions group */}
                   {classification.newSessions.length > 0 && (
                     <>
                       <Text size="3" weight="bold">
@@ -360,17 +305,12 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
                             checkbox
                             checked={newSessionSelectedIds.has(session.id)}
                             onToggle={() => toggleNewSession(session.id)}
-                            formatDate={formatDate}
-                            getGroupLabel={getGroupLabel}
-                            getTabLabel={getTabLabel}
-                            getTabCount={getTabCount}
                           />
                         ))}
                       </Flex>
                     </>
                   )}
 
-                  {/* Conflicting sessions group */}
                   {classification.conflictingSessions.length > 0 && (
                     <>
                       {classification.newSessions.length > 0 && <Separator size="4" />}
@@ -385,34 +325,20 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
                           onValueChange={(v: string) => setConflictMode(v as ConflictMode)}
                           size="1"
                         >
-                          <SegmentedControl.Item value="overwrite">
-                            {getMessage('conflictModeOverwrite')}
-                          </SegmentedControl.Item>
-                          <SegmentedControl.Item value="duplicate">
-                            {getMessage('conflictModeDuplicate')}
-                          </SegmentedControl.Item>
-                          <SegmentedControl.Item value="ignore">
-                            {getMessage('conflictModeIgnore')}
-                          </SegmentedControl.Item>
+                          <SegmentedControl.Item value="overwrite">{getMessage('conflictModeOverwrite')}</SegmentedControl.Item>
+                          <SegmentedControl.Item value="duplicate">{getMessage('conflictModeDuplicate')}</SegmentedControl.Item>
+                          <SegmentedControl.Item value="ignore">{getMessage('conflictModeIgnore')}</SegmentedControl.Item>
                         </SegmentedControl.Root>
                       </Flex>
 
                       <Flex direction="column" gap="2">
                         {classification.conflictingSessions.map(conflict => (
-                          <ConflictSessionRow
-                            key={conflict.imported.id}
-                            conflict={conflict}
-                            formatDate={formatDate}
-                            getGroupLabel={getGroupLabel}
-                            getTabLabel={getTabLabel}
-                            getTabCount={getTabCount}
-                          />
+                          <ConflictSessionRow key={conflict.imported.id} conflict={conflict} />
                         ))}
                       </Flex>
                     </>
                   )}
 
-                  {/* Identical sessions group */}
                   {classification.identicalSessions.length > 0 && (
                     <>
                       {(classification.newSessions.length > 0 || classification.conflictingSessions.length > 0) && (
@@ -428,10 +354,6 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
                             session={session}
                             dimmed
                             statusBadge={getMessage('alreadyExists')}
-                            formatDate={formatDate}
-                            getGroupLabel={getGroupLabel}
-                            getTabLabel={getTabLabel}
-                            getTabCount={getTabCount}
                           />
                         ))}
                       </Flex>
@@ -446,12 +368,8 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
 
               {hasConflicts && conflictMode === 'overwrite' && (
                 <Callout.Root color="orange" variant="soft" mt="3">
-                  <Callout.Icon>
-                    <AlertTriangle size={16} />
-                  </Callout.Icon>
-                  <Callout.Text>
-                    {getMessage('sessionImportOverwriteWarning')}
-                  </Callout.Text>
+                  <Callout.Icon><AlertTriangle size={16} /></Callout.Icon>
+                  <Callout.Text>{getMessage('sessionImportOverwriteWarning')}</Callout.Text>
                 </Callout.Root>
               )}
             </Box>
@@ -459,14 +377,11 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
 
           <Separator size="4" mt="4" style={{ opacity: 0.3 }} />
 
-          {/* Footer */}
           <Flex gap="3" justify="end" mt="3">
             {step === 0 && (
               <>
                 <Dialog.Close>
-                  <Button variant="soft" color="gray">
-                    {getMessage('cancel')}
-                  </Button>
+                  <Button variant="soft" color="gray">{getMessage('cancel')}</Button>
                 </Dialog.Close>
                 <Button onClick={goToStep1} disabled={!parsedSessions}>
                   {getMessage('next')}
@@ -487,248 +402,5 @@ export function ImportSessionsWizard({ open, onOpenChange }: ImportSessionsWizar
         </Dialog.Content>
       </Dialog.Root>
     </SessionsTheme>
-  );
-}
-
-// --- Sub-components ---
-
-interface SessionRowProps {
-  session: Session;
-  checkbox?: boolean;
-  checked?: boolean;
-  onToggle?: () => void;
-  dimmed?: boolean;
-  statusBadge?: string;
-  formatDate: (iso: string) => string;
-  getGroupLabel: (count: number) => string;
-  getTabLabel: (count: number) => string;
-  getTabCount: (session: Session) => number;
-}
-
-function SessionRow({
-  session,
-  checkbox,
-  checked,
-  onToggle,
-  dimmed,
-  statusBadge,
-  formatDate,
-  getGroupLabel,
-  getTabLabel,
-  getTabCount,
-}: SessionRowProps) {
-  const tabCount = getTabCount(session);
-  const groupCount = session.groups.length;
-
-  return (
-    <Flex
-      align="center"
-      gap="3"
-      p="2"
-      style={{
-        borderRadius: 'var(--radius-2)',
-        backgroundColor: 'var(--gray-a2)',
-        opacity: dimmed ? 0.6 : 1,
-      }}
-    >
-      {checkbox && (
-        <Checkbox
-          checked={checked}
-          onCheckedChange={onToggle}
-          aria-label={session.name}
-        />
-      )}
-      <Flex direction="column" gap="1" style={{ flex: 1 }}>
-        <Text size="2" weight="medium">{session.name}</Text>
-        <Text size="1" color="gray">
-          {formatDate(session.createdAt)} · {getGroupLabel(groupCount)} · {getTabLabel(tabCount)}
-        </Text>
-      </Flex>
-      {session.isPinned && (
-        <Badge color="indigo" variant="soft" size="1">
-          {getMessage('pinnedBadge')}
-        </Badge>
-      )}
-      {statusBadge && (
-        <Badge color="gray" variant="outline" size="1">
-          {statusBadge}
-        </Badge>
-      )}
-    </Flex>
-  );
-}
-
-interface ConflictSessionRowProps {
-  conflict: ConflictingSession;
-  formatDate: (iso: string) => string;
-  getGroupLabel: (count: number) => string;
-  getTabLabel: (count: number) => string;
-  getTabCount: (session: Session) => number;
-}
-
-function ConflictSessionRow({
-  conflict,
-  formatDate,
-  getGroupLabel,
-  getTabLabel,
-  getTabCount,
-}: ConflictSessionRowProps) {
-  const session = conflict.imported;
-  const tabCount = getTabCount(session);
-  const groupCount = session.groups.length;
-
-  return (
-    <Flex
-      align="center"
-      gap="3"
-      p="2"
-      style={{
-        borderRadius: 'var(--radius-2)',
-        backgroundColor: 'var(--orange-a2)',
-      }}
-    >
-      <AlertTriangle size={16} style={{ color: 'var(--orange-9)', flexShrink: 0 }} aria-hidden="true" />
-      <Flex direction="column" gap="1" style={{ flex: 1 }}>
-        <Text size="2" weight="medium">{session.name}</Text>
-        <Text size="1" color="gray">
-          {formatDate(session.createdAt)} · {getGroupLabel(groupCount)} · {getTabLabel(tabCount)}
-        </Text>
-      </Flex>
-      {session.isPinned && (
-        <Badge color="indigo" variant="soft" size="1">
-          {getMessage('pinnedBadge')}
-        </Badge>
-      )}
-      <Popover.Root>
-        <Popover.Trigger>
-          <Button variant="ghost" size="1" aria-label={getMessage('viewDiff')} title={getMessage('viewDiff')}>
-            <Eye size={14} aria-hidden="true" />
-          </Button>
-        </Popover.Trigger>
-        <Popover.Content style={{ maxWidth: 380 }}>
-          <Text size="3" weight="bold" mb="2">
-            {getMessage('differences')} — {session.name}
-          </Text>
-          <SessionDiffView diff={conflict.diff} />
-        </Popover.Content>
-      </Popover.Root>
-    </Flex>
-  );
-}
-
-interface SessionDiffViewProps {
-  diff: SessionDiff;
-}
-
-function SessionDiffView({ diff }: SessionDiffViewProps) {
-  const hasContent =
-    diff.isPinned !== undefined ||
-    diff.categoryId !== undefined ||
-    diff.groupsAdded.length > 0 ||
-    diff.groupsRemoved.length > 0 ||
-    diff.groupsChanged.length > 0 ||
-    diff.ungroupedTabsAdded.length > 0 ||
-    diff.ungroupedTabsRemoved.length > 0;
-
-  if (!hasContent) return null;
-
-  return (
-    <Flex direction="column" gap="3">
-      {diff.isPinned !== undefined && (
-        <Box>
-          <Text size="2" weight="medium" mb="1">isPinned</Text>
-          <Flex direction="column" gap="1">
-            <Flex align="center" gap="2">
-              <Badge color="red" variant="soft" size="1">{getMessage('currentValue')}</Badge>
-              <Text size="1">{String(diff.isPinned.current)}</Text>
-            </Flex>
-            <Flex align="center" gap="2">
-              <Badge color="green" variant="soft" size="1">{getMessage('importedValue')}</Badge>
-              <Text size="1">{String(diff.isPinned.imported)}</Text>
-            </Flex>
-          </Flex>
-        </Box>
-      )}
-
-      {diff.categoryId !== undefined && (
-        <Box>
-          <Text size="2" weight="medium" mb="1">categoryId</Text>
-          <Flex direction="column" gap="1">
-            <Flex align="center" gap="2">
-              <Badge color="red" variant="soft" size="1">{getMessage('currentValue')}</Badge>
-              <Text size="1">{String(diff.categoryId.current ?? 'null')}</Text>
-            </Flex>
-            <Flex align="center" gap="2">
-              <Badge color="green" variant="soft" size="1">{getMessage('importedValue')}</Badge>
-              <Text size="1">{String(diff.categoryId.imported ?? 'null')}</Text>
-            </Flex>
-          </Flex>
-        </Box>
-      )}
-
-      {diff.groupsAdded.length > 0 && (
-        <Box>
-          <Text size="2" weight="medium" mb="1">Groups added</Text>
-          {diff.groupsAdded.map(title => (
-            <Text key={title} size="1" color="green" as="p">{title}</Text>
-          ))}
-        </Box>
-      )}
-
-      {diff.groupsRemoved.length > 0 && (
-        <Box>
-          <Text size="2" weight="medium" mb="1">Groups removed</Text>
-          {diff.groupsRemoved.map(title => (
-            <Text key={title} size="1" color="red" as="p">{title}</Text>
-          ))}
-        </Box>
-      )}
-
-      {diff.groupsChanged.map(groupDiff => (
-        <GroupDiffView key={groupDiff.title} groupDiff={groupDiff} />
-      ))}
-
-      {(diff.ungroupedTabsAdded.length > 0 || diff.ungroupedTabsRemoved.length > 0) && (
-        <Box>
-          <Text size="2" weight="medium" mb="1">Ungrouped tabs</Text>
-          {diff.ungroupedTabsAdded.map(url => (
-            <Text key={url} size="1" color="green" as="p" style={{ wordBreak: 'break-all' }}>+ {url}</Text>
-          ))}
-          {diff.ungroupedTabsRemoved.map(url => (
-            <Text key={url} size="1" color="red" as="p" style={{ wordBreak: 'break-all' }}>- {url}</Text>
-          ))}
-        </Box>
-      )}
-    </Flex>
-  );
-}
-
-interface GroupDiffViewProps {
-  groupDiff: GroupDiff;
-}
-
-function GroupDiffView({ groupDiff }: GroupDiffViewProps) {
-  return (
-    <Box>
-      <Text size="2" weight="medium" mb="1">Group: {groupDiff.title}</Text>
-      {groupDiff.colorChanged && (
-        <Flex direction="column" gap="1" mb="1">
-          <Flex align="center" gap="2">
-            <Badge color="red" variant="soft" size="1">{getMessage('currentValue')}</Badge>
-            <Text size="1">{groupDiff.colorChanged.current}</Text>
-          </Flex>
-          <Flex align="center" gap="2">
-            <Badge color="green" variant="soft" size="1">{getMessage('importedValue')}</Badge>
-            <Text size="1">{groupDiff.colorChanged.imported}</Text>
-          </Flex>
-        </Flex>
-      )}
-      {groupDiff.tabsAdded.map(url => (
-        <Text key={url} size="1" color="green" as="p" style={{ wordBreak: 'break-all' }}>+ {url}</Text>
-      ))}
-      {groupDiff.tabsRemoved.map(url => (
-        <Text key={url} size="1" color="red" as="p" style={{ wordBreak: 'break-all' }}>- {url}</Text>
-      ))}
-    </Box>
   );
 }
