+++ src/pages/options.tsx
@@ -1,5 +1,5 @@
 // options/options.ts
-import React, { useState, useEffect, useCallback } from 'react';
+import React, { useCallback } from 'react';
 import { createRoot } from 'react-dom/client';
 import { browser } from 'wxt/browser';
 import { Theme } from '@radix-ui/themes';
@@ -7,274 +7,61 @@ import { ThemeProvider } from 'next-themes';
 
 import { useSyncedSettings } from '../hooks/useSyncedSettings.js';
 import { useStatistics } from '../hooks/useStatistics.js';
-import { generateUUID, isValidDomain, isValidRegex } from '../utils/utils';
+import { useDeepLinking } from '../hooks/useDeepLinking.js';
 import { getMessage } from '../utils/i18n';
 const version = browser.runtime.getManifest().version;
 
 import { Sidebar } from '../components/UI/Sidebar/Sidebar';
 import type { SidebarItem } from '../components/UI/Sidebar/Sidebar';
-import { ThemeToggle } from '../components/UI/ThemeToggle/ThemeToggle';
-import { Flex, Text } from '@radix-ui/themes';
+import { OptionsHeader, OptionsHeaderCollapsed } from '../components/UI/OptionsLayout/OptionsHeader';
+import { OptionsFooter, OptionsFooterCollapsed } from '../components/UI/OptionsLayout/OptionsFooter';
+import { useState } from 'react';
 import { DomainRulesPage } from './DomainRulesPage';
 import { StatisticsPage } from './StatisticsPage';
 import { SettingsPage } from '../components/UI/SettingsPage/SettingsPage';
 import { ImportExportPage } from '../components/UI/ImportExportPage/ImportExportPage';
 import { ConfirmDialog } from '../components/UI/ConfirmDialog/ConfirmDialog';
-import { Shield, FileText, BarChart3, Settings, Github, Archive } from 'lucide-react';
+import { Shield, FileText, BarChart3, Settings, Archive } from 'lucide-react';
 import { FEATURE_BASE_COLORS } from '../utils/themeConstants';
 import { SessionsPage } from './SessionsPage';
-import type { SyncSettings, DomainRuleSettings } from '../types/syncSettings';
+import type { DomainRuleSettings } from '../types/syncSettings';
 
 (() => {
 
-// --- Fonctions Utilitaires ---
-interface TooltipProps {
-    textKey: string;
-    children: React.ReactNode;
-}
-
-function Tooltip({ textKey, children }: TooltipProps) {
-    return (
-        <div className="tooltip-container">
-            {children}
-            <span className="tooltip-text">{getMessage(textKey)}</span>
-        </div>
-    );
-}
-
-// --- Composant Principal ---
 function OptionsContent() {
     const { settings, updateSettings } = useSyncedSettings();
     const { statistics: stats, resetStatistics } = useStatistics();
-    const [currentTab, setCurrentTab] = useState<string>('rules');
-    
+    const {
+        currentTab, setCurrentTab,
+        openSnapshotWizard, setOpenSnapshotWizard,
+        snapshotGroupId, setSnapshotGroupId,
+    } = useDeepLinking();
+
     const [sidebarCollapsed, setSidebarCollapsed] = useState<boolean>(false);
     const [resetConfirmOpen, setResetConfirmOpen] = useState(false);
-    const [openSnapshotWizard, setOpenSnapshotWizard] = useState(false);
-    const [snapshotGroupId, setSnapshotGroupId] = useState<number | null>(null);
-
-    // Deep linking: navigate to the right section on mount and on hash changes
-    // (hashchange fires when an existing tab's URL hash is updated externally)
-    useEffect(() => {
-        const validSections = ['rules', 'importexport', 'sessions', 'stats', 'settings'];
-
-        function handleHash() {
-            const hash = window.location.hash; // e.g. '#sessions?action=snapshot'
-            if (!hash.startsWith('#')) return;
-            const questionMark = hash.indexOf('?');
-            const section = questionMark === -1 ? hash.slice(1) : hash.slice(1, questionMark);
-            if (!validSections.includes(section)) return;
-            setCurrentTab(section);
-            if (section === 'sessions' && questionMark !== -1) {
-                const params = new URLSearchParams(hash.slice(questionMark + 1));
-                if (params.get('action') === 'snapshot') {
-                    setOpenSnapshotWizard(true);
-                    const groupIdParam = params.get('groupId');
-                    setSnapshotGroupId(groupIdParam ? parseInt(groupIdParam, 10) : null);
-                }
-            }
-        }
-
-        handleHash();
-        window.addEventListener('hashchange', handleHash);
-        return () => window.removeEventListener('hashchange', handleHash);
-    }, []);
-
-    // Storage handled by hooks
-
-    // --- Gestionnaires ---
-    const updateSetting = useCallback((key: keyof SyncSettings, value: any) => {
-        updateSettings({ [key]: value });
-    }, [updateSettings]);
 
     const updateRules = useCallback((newRules: DomainRuleSettings) => {
         updateSettings({ domainRules: newRules });
     }, [updateSettings]);
 
-    
-
-
-     const handleResetStats = useCallback(() => {
-         setResetConfirmOpen(true);
-    }, []);
+    const handleResetStats = useCallback(() => setResetConfirmOpen(true), []);
 
     const handleTabChange = useCallback((tab: string) => {
         window.location.hash = tab;
         setCurrentTab(tab);
-    }, []);
+    }, [setCurrentTab]);
 
-    // Sidebar items configuration with themed accent colors
     const sidebarItems: SidebarItem[] = [
-        {
-            id: 'rules',
-            label: getMessage('domainRulesTab'),
-            icon: Shield as any,
-            accentColor: FEATURE_BASE_COLORS.DOMAIN_RULES,
-        },
-        {
-            id: 'sessions',
-            label: getMessage('sessionsTab'),
-            icon: Archive as any,
-            accentColor: FEATURE_BASE_COLORS.SESSIONS,
-        },
-        {
-            id: 'importexport',
-            label: getMessage('importExportTab'),
-            icon: FileText as any,
-            accentColor: FEATURE_BASE_COLORS.IMPORT, // Utilise la couleur Import pour l'onglet combiné
-        },
-        {
-            id: 'stats',
-            label: getMessage('statisticsTab'),
-            icon: BarChart3 as any,
-            accentColor: FEATURE_BASE_COLORS.STATISTICS,
-        },
-        {
-            id: 'settings',
-            label: getMessage('settingsTab'),
-            icon: Settings as any,
-            accentColor: FEATURE_BASE_COLORS.SETTINGS,
-        },
+        { id: 'rules', label: getMessage('domainRulesTab'), icon: Shield as any, accentColor: FEATURE_BASE_COLORS.DOMAIN_RULES },
+        { id: 'sessions', label: getMessage('sessionsTab'), icon: Archive as any, accentColor: FEATURE_BASE_COLORS.SESSIONS },
+        { id: 'importexport', label: getMessage('importExportTab'), icon: FileText as any, accentColor: FEATURE_BASE_COLORS.IMPORT },
+        { id: 'stats', label: getMessage('statisticsTab'), icon: BarChart3 as any, accentColor: FEATURE_BASE_COLORS.STATISTICS },
+        { id: 'settings', label: getMessage('settingsTab'), icon: Settings as any, accentColor: FEATURE_BASE_COLORS.SETTINGS },
     ];
 
-
-    // --- Rendu ---
     if (!settings) {
         return <p>Chargement...</p>;
     }
-
-    // Contenu du header pour la sidebar
-    const headerContent = (
-        <Flex align="center" gap="3" style={{ width: '100%', paddingRight: '64px', position: 'relative' }}>
-            <img 
-                src="/icons/icon48.png" 
-                alt="SmartTab Organizer" 
-                style={{ 
-                    width: '32px', 
-                    height: '32px',
-                    flexShrink: 0
-                }} 
-            />
-            <Flex direction="column" gap="0" style={{ lineHeight: '1.2', flex: 1 }}>
-                <Flex align="center" gap="2">
-                    <Text size="3" weight="bold" style={{ color: 'var(--gray-12)' }}>
-                        SmartTab
-                    </Text>
-                    <Text size="1" style={{ color: 'var(--gray-11)' }}>
-                        (v{version})
-                    </Text>
-                </Flex>
-                <Text size="3" weight="bold" style={{ color: 'var(--gray-12)' }}>
-                    Organizer
-                </Text>
-            </Flex>
-            <Flex align="center" style={{ position: 'absolute', right: '8px' }}>
-                <ThemeToggle />
-            </Flex>
-        </Flex>
-    );
-
-    const headerCollapsedContent = (
-        <Flex align="center" justify="center" style={{ width: '100%' }}>
-            <img 
-                src="/icons/icon48.png" 
-                alt="SmartTab Organizer" 
-                style={{ 
-                    width: '32px', 
-                    height: '32px'
-                }} 
-            />
-        </Flex>
-    );
-
-    // Custom SidebarFooter content
-    const customFooterContent = (
-        <Flex 
-            align="center" 
-            gap="3" 
-            style={{ 
-                padding: '12px 16px',
-                cursor: 'pointer',
-                borderRadius: '4px',
-                transition: 'background-color 0.2s ease',
-            }}
-            onClick={() => browser.tabs.create({ url: 'https://github.com/EspritVorace/smart-tab-organizer' })}
-            onKeyDown={(e) => {
-                if (e.key === 'Enter' || e.key === ' ') {
-                    e.preventDefault();
-                    browser.tabs.create({ url: 'https://github.com/EspritVorace/smart-tab-organizer' });
-                }
-            }}
-            tabIndex={0}
-            role="button"
-            aria-label="Visiter le profil GitHub d'EspritVorace"
-            className="custom-footer-expanded"
-        >
-            <img 
-                src="/icons/ev.png" 
-                alt="EspritVorace" 
-                style={{ 
-                    width: '24px', 
-                    height: '24px',
-                    borderRadius: '50%',
-                    flexShrink: 0
-                }} 
-            />
-            <Flex align="center" gap="2">
-                <Text 
-                    size="2" 
-                    weight="medium" 
-                    style={{ 
-                        color: 'var(--gray-11)',
-                        textDecoration: 'none'
-                    }}
-                >
-                    EspritVorace
-                </Text>
-                <Github 
-                    size={16} 
-                    style={{ 
-                        color: 'var(--gray-11)',
-                        flexShrink: 0
-                    }} 
-                />
-            </Flex>
-        </Flex>
-    );
-
-    const customFooterCollapsedContent = (
-        <Flex 
-            align="center" 
-            justify="center" 
-            style={{ 
-                padding: '12px',
-                cursor: 'pointer',
-                borderRadius: '4px',
-                transition: 'background-color 0.2s ease',
-            }}
-            onClick={() => browser.tabs.create({ url: 'https://github.com/EspritVorace/smart-tab-organizer' })}
-            onKeyDown={(e) => {
-                if (e.key === 'Enter' || e.key === ' ') {
-                    e.preventDefault();
-                    browser.tabs.create({ url: 'https://github.com/EspritVorace/smart-tab-organizer' });
-                }
-            }}
-            tabIndex={0}
-            role="button"
-            aria-label="Visiter le profil GitHub d'EspritVorace"
-            className="custom-footer-collapsed"
-        >
-            <img 
-                src="/icons/ev.png" 
-                alt="EspritVorace" 
-                style={{ 
-                    width: '24px', 
-                    height: '24px',
-                    borderRadius: '50%'
-                }} 
-            />
-        </Flex>
-    );
 
     return (
         <div id="options-inner" style={{ display: 'flex', height: '100vh' }}>
@@ -284,11 +71,11 @@ function OptionsContent() {
                 activeItem={currentTab}
                 onItemClick={handleTabChange}
                 items={sidebarItems}
-                headerContent={headerContent}
-                headerCollapsedContent={headerCollapsedContent}
+                headerContent={<OptionsHeader version={version} />}
+                headerCollapsedContent={<OptionsHeaderCollapsed />}
                 showFooter={true}
-                footerContent={customFooterContent}
-                footerCollapsedContent={customFooterCollapsedContent}
+                footerContent={<OptionsFooter />}
+                footerCollapsedContent={<OptionsFooterCollapsed />}
             />
             <div style={{ flex: 1, display: 'flex', flexDirection: 'column' }}>
                 <main style={{ flex: 1, overflow: 'auto', padding: '20px' }}>
@@ -329,17 +116,11 @@ function OptionsContent() {
             />
         </div>
     );
-
 }
 
 function OptionsApp() {
     return (
-        <ThemeProvider
-            attribute="class"
-            defaultTheme="system"
-            enableSystem
-            disableTransitionOnChange
-        >
+        <ThemeProvider attribute="class" defaultTheme="system" enableSystem disableTransitionOnChange>
             <Theme>
                 <OptionsContent />
             </Theme>
